feat: Model unidirectional balances
These changes allow for a more natural accessing of the sender and
receiver balances. The state channels are unidirectional. A  channel
begins with an initial `sender_balance` that is wholly in the possession
of the sender. Each valid transfer can only decrease the
`sender_balance` and increase the `receiver_balance` by the same amount.

The `nonce` is incremented by 1 for each valid transfer. Zero-value
transfer amounts are disallowed.